+++ src/components/Comment/components/CommentEllipsisButton.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import CommentContextMenu from '@components/Common/ContextMenu/components/CommentContextMenu';
 import Ellipsis from '@components/Common/Icons/Ellipsis';
 import { useTheme } from 'tamagui';
-import { Pressable } from 'react-native';
 
 interface IProps {
   itemId: number;
@@ -15,9 +14,7 @@ export default function CommentEllipsisButton({
 
   return (
     <CommentContextMenu itemId={itemId}>
-      <Pressable hitSlop={120}>
-        <Ellipsis size={16} color={theme.accent.val} />
-      </Pressable>
+      <Ellipsis size={16} color={theme.accent.val} />
     </CommentContextMenu>
   );
 }