+++ src/components/Common/PostCard/PostUserLabel.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback, useMemo } from 'react';
 import { useSettingsStore } from '@src/state';
 import { Text, XStack } from 'tamagui';
-import { Pressable } from 'react-native';
 import { useNavigation } from '@react-navigation/core';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import { createName } from '@helpers/text';
@@ -41,16 +40,14 @@ function PostUserLabel({
   }, [name]);
 
   return (
-    <Pressable onPress={onPress} hitSlop={5}>
-      <XStack space="$2" alignItems="center">
-        {showAvatar && <UserIcon userIcon={userIcon} />}
-        <Text color="$secondary" fontSize="$2">
-          {name}
-        </Text>
-        {isAdmin === true && <AdminBadge />}
-        {isSelf === true && <AdminBadge />}
-      </XStack>
-    </Pressable>
+    <XStack space="$2" alignItems="center" width="auto">
+      {showAvatar && <UserIcon userIcon={userIcon} />}
+      <Text color="$secondary" fontSize="$2" onPress={onPress} hitSlop={3}>
+        {name}
+      </Text>
+      {isAdmin === true && <AdminBadge />}
+      {isSelf === true && <AdminBadge />}
+    </XStack>
   );
 }