+++ src/components/Post/components/PostContextButton.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useTheme, XStack } from 'tamagui';
 import PostContextMenu from '@components/Common/ContextMenu/components/PostContextMenu';
-import { Pressable } from 'react-native';
 import Ellipsis from '@components/Common/Icons/Ellipsis';
 
 interface IProps {
@@ -14,9 +13,7 @@ function PostContextButton({ itemId }: IProps): React.JSX.Element {
   return (
     <XStack ml="auto" p="$1" pr="$2" space="$3">
       <PostContextMenu itemId={itemId}>
-        <Pressable hitSlop={120}>
-          <Ellipsis size={18} color={theme.accent.val} />
-        </Pressable>
+        <Ellipsis size={18} color={theme.accent.val} />
       </PostContextMenu>
     </XStack>
   );