+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Drawer not going to community
+- The hitbox of an image is no longer 100% of the width of the screen
 
 ## [Version 1.0 (58)] - 2023-10-27T10:59:52Z